Seperate wlr-screencopy-unstable-v1 code from generic screencast code
